Remove AbstractTieredStorageAssetManagerTest
